+++ src/roslibpy/ros.py
@@ -4,6 +4,7 @@ import logging
 import threading
 
 from . import Message
+from . import Param
 from . import Service
 from . import ServiceRequest
 from .comm import RosBridgeClientFactory
@@ -11,10 +12,18 @@ from .comm import RosBridgeClientFactory
 __all__ = ['Ros']
 
 LOGGER = logging.getLogger('roslibpy')
+CONNECTION_TIMEOUT = 10
+ROSAPI_TIMEOUT = 3
 
 
 class Ros(object):
-    """Connection manager to ROS server."""
+    """Connection manager to ROS server.
+
+    Args:
+        host (:obj:`str`): Name or IP address of the ROS bridge host, e.g. ``127.0.0.1``.
+        port (:obj:`int`): ROS bridge port, e.g. ``9090``.
+        is_secure (:obj:`bool`): ``True`` to use a secure web sockets connection, otherwise ``False``.
+    """
 
     def __init__(self, host, port=None, is_secure=False):
         self._id_counter = 0
@@ -60,12 +69,13 @@ class Ros(object):
         """Disconnect from ROS master."""
         if self.is_connected:
             def _wrapper_callback(proto):
+                self.emit('closing')
                 proto.send_close()
                 return proto
 
             self.factory.on_ready(_wrapper_callback)
 
-    def run(self, timeout=None):
+    def run(self, timeout=CONNECTION_TIMEOUT):
         """Kick-starts a non-blocking event loop.
 
         Args:
@@ -177,7 +187,58 @@ class Ros(object):
 
         self.factory.on_ready(_send_internal)
 
-    def send_service_request(self, message, callback, errback):
+    def blocking_call_from_thread(self, callback, timeout):
+        """Call the given function from a thread, and wait for the result synchronously
+        for as long as the timeout will allow.
+
+        Args:
+            callback: Callable function to be invoked from the thread.
+            timeout (:obj: int): Number of seconds to wait for the response before
+            raising an exception.
+
+        Returns:
+            The results from the callback, or a timeout exception.
+        """
+        return self.factory.manager.blocking_call_from_thread(callback, timeout)
+
+    def get_service_request_callback(self, message):
+        """Get the callback which, when called, sends the service request.
+
+        Args:
+            message (:class:`.Message`): ROS Bridge Message containing the request.
+
+        Returns:
+            A callable which makes the service request.
+        """
+        def get_call_results(result_placeholder):
+
+            inner_callback = self.factory.manager.get_inner_callback(result_placeholder)
+
+            inner_errback = self.factory.manager.get_inner_errback(result_placeholder)
+
+            self.call_async_service(message, inner_callback, inner_errback)
+
+            return result_placeholder
+
+        return get_call_results
+
+    def call_sync_service(self, message, timeout):
+        """Send a blocking service request to the ROS Master once the connection is established,
+        waiting for the result to be return.
+
+        If a connection to ROS is already available, the request is sent immediately.
+
+        Args:
+            message (:class:`.Message`): ROS Bridge Message containing the request.
+            timeout (:obj: int): Number of seconds to wait for the response before
+            raising an exception.
+        Returns:
+            Either returns the service request results or raises a timeout exception.
+        """
+        get_call_results = self.get_service_request_callback(message)
+        return self.blocking_call_from_thread(get_call_results, timeout)
+
+    def call_async_service(self, message, callback, errback):
         """Send a service request to the ROS Master once the connection is established.
 
         If a connection to ROS is already available, the request is sent immediately.
@@ -202,6 +263,242 @@ class Ros(object):
 
         self.send_on_ready(level_message)
 
+    def get_topics(self, callback=None, errback=None):
+        """Retrieve list of topics in ROS.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+
+        Returns:
+            list: List of topics if blocking, otherwise ``None``.
+        """
+        service = Service(self, '/rosapi/topics', 'rosapi/Topics')
+
+        result = service.call(ServiceRequest(), callback,
+                              errback, timeout=ROSAPI_TIMEOUT)
+
+        if callback:
+            return
+
+        assert 'topics' in result
+        return result['topics']
+
+    def get_topic_type(self, topic, callback=None, errback=None):
+        """Retrieve the type of a topic in ROS.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+
+        Returns:
+            str: Topic type if blocking, otherwise ``None``.
+        """
+        service = Service(self, '/rosapi/topic_type',
+                          'rosapi/TopicType')
+
+        result = service.call(ServiceRequest({'topic': topic}),
+                              callback, errback, timeout=ROSAPI_TIMEOUT)
+
+        if callback:
+            return
+
+        assert 'type' in result
+        return result['type']
+
+    def get_topics_for_type(self, topic_type, callback=None, errback=None):
+        """Retrieve list of topics in ROS matching the specified type.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+
+        Returns:
+            list: List of topics matching the specified type if blocking, otherwise ``None``.
+        """
+        service = Service(self, '/rosapi/topics_for_type',
+                          'rosapi/TopicsForType')
+
+        result = service.call(ServiceRequest({'type': topic_type}),
+                              callback, errback, timeout=ROSAPI_TIMEOUT)
+
+        if callback:
+            return
+
+        assert 'topics' in result
+        return result['topics']
+
+    def get_services(self, callback=None, errback=None):
+        """Retrieve list of active service names in ROS.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+
+        Returns:
+            list: List of services if blocking, otherwise ``None``.
+        """
+        service = Service(self, '/rosapi/services',
+                          'rosapi/Services')
+
+        result = service.call(ServiceRequest(), callback,
+                              errback, timeout=ROSAPI_TIMEOUT)
+
+        if callback:
+            return
+
+        assert 'services' in result
+        return result['services']
+
+    def get_service_type(self, service_name, callback=None, errback=None):
+        """Retrieve the type of a service in ROS.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+
+        Returns:
+            str: Service type if blocking, otherwise ``None``.
+        """
+        service = Service(self, '/rosapi/service_type',
+                          'rosapi/ServiceType')
+
+        result = service.call(ServiceRequest(
+            {'service': service_name}), callback, errback, timeout=ROSAPI_TIMEOUT)
+
+        if callback:
+            return
+
+        assert 'type' in result
+        return result['type']
+
+    def get_services_for_type(self, service_type, callback=None, errback=None):
+        """Retrieve list of services in ROS matching the specified type.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+
+        Returns:
+            list: List of services matching the specified type if blocking, otherwise ``None``.
+        """
+        service = Service(self, '/rosapi/services_for_type',
+                          'rosapi/ServicesForType')
+
+        result = service.call(ServiceRequest({'type': service_type}),
+                              callback, errback, timeout=ROSAPI_TIMEOUT)
+
+        if callback:
+            return
+
+        assert 'services' in result
+        return result['services']
+
+    def get_service_request_details(self, type, callback=None, errback=None):
+        """Retrieve details of a ROS Service Request.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+
+        Returns:
+            Service Request details if blocking, otherwise ``None``.
+        """
+        service = Service(self, '/rosapi/service_request_details',
+                          'rosapi/ServiceRequestDetails')
+
+        result = service.call(ServiceRequest({'type': type}),
+                              callback, errback, timeout=ROSAPI_TIMEOUT)
+
+        if callback:
+            return
+
+        return result
+
+    def get_service_response_details(self, type, callback=None, errback=None):
+        """Retrieve details of a ROS Service Response.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+
+        Returns:
+            Service Response details if blocking, otherwise ``None``.
+        """
+        service = Service(self, '/rosapi/service_response_details',
+                          'rosapi/ServiceResponseDetails')
+
+        result = service.call(ServiceRequest({'type': type}),
+                              callback, errback, timeout=ROSAPI_TIMEOUT)
+
+        if callback:
+            return
+
+        return result
+
+    def get_message_details(self, message_type, callback=None, errback=None):
+        """Retrieve details of a message type in ROS.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+
+        Returns:
+            Message type details if blocking, otherwise ``None``.
+        """
+        service = Service(self, '/rosapi/message_details',
+                          'rosapi/MessageDetails')
+
+        result = service.call(ServiceRequest(
+            {'type': message_type}), callback, errback, timeout=ROSAPI_TIMEOUT)
+
+        if callback:
+            return
+
+        return result
+
+    def get_params(self, callback=None, errback=None):
+        """Retrieve list of param names from the ROS Parameter Server.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+
+        Returns:
+            list: List of parameters if blocking, otherwise ``None``.
+        """
+        service = Service(self, '/rosapi/get_param_names',
+                          'rosapi/GetParamNames')
+
+        result = service.call(ServiceRequest(), callback,
+                              errback, timeout=ROSAPI_TIMEOUT)
+
+        if callback:
+            return
+
+        assert 'names' in result
+        return result['names']
+
+    def get_param(self, name, callback=None, errback=None):
+        """Get the value of a parameter from the ROS Parameter Server.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+
+        Returns:
+            Parameter value if blocking, otherwise ``None``.
+        """
+        param = Param(self, name)
+        return param.get(callback, errback, timeout=ROSAPI_TIMEOUT)
+
+    def set_param(self, name, value, callback=None, errback=None):
+        """Set the value of a parameter from the ROS Parameter Server.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+        """
+        param = Param(self, name)
+        param.set(value, callback, errback, timeout=ROSAPI_TIMEOUT)
+
+    def delete_param(self, name, callback=None, errback=None):
+        """Delete parameter from the ROS Parameter Server.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+        """
+        param = Param(self, name)
+        param.delete(callback, errback, timeout=ROSAPI_TIMEOUT)
+
     def get_action_servers(self, callback, errback=None):
         """Retrieve list of action servers in ROS."""
         service = Service(self, '/rosapi/action_servers',
@@ -209,91 +506,43 @@ class Ros(object):
 
         service.call(ServiceRequest(), callback, errback)
 
-    def get_topics(self, callback, errback=None):
-        """Retrieve list of topics in ROS."""
-        service = Service(self, '/rosapi/topics',
-                          'rosapi/Topics')
+    def get_nodes(self, callback=None, errback=None):
+        """Retrieve list of active node names in ROS.
 
-        service.call(ServiceRequest(), callback, errback)
-
-    def get_topics_for_type(self, topic_type, callback, errback=None):
-        """Retrieve list of topics in ROS matching the specified type."""
-        service = Service(self, '/rosapi/topics_for_type',
-                          'rosapi/TopicsForType')
-
-        service.call(ServiceRequest({'type': topic_type}), callback, errback)
-
-    def get_services(self, callback, errback=None):
-        """Retrieve list of active service names in ROS."""
-        service = Service(self, '/rosapi/services',
-                          'rosapi/Services')
-
-        service.call(ServiceRequest(), callback, errback)
-
-    def get_services_for_type(self, service_type, callback, errback=None):
-        """Retrieve list of services in ROS matching the specified type."""
-        service = Service(self, '/rosapi/services_for_type',
-                          'rosapi/ServicesForType')
-
-        service.call(ServiceRequest({'type': service_type}), callback, errback)
-
-    def get_service_request_details(self, type, callback, errback=None):
-        """Retrieve details of a ROS Service Request."""
-        service = Service(self, '/rosapi/service_request_details',
-                          'rosapi/ServiceRequestDetails')
-
-        service.call(ServiceRequest({'type': type}), callback, errback)
-
-    def get_service_response_details(self, type, callback, errback=None):
-        """Retrieve details of a ROS Service Response."""
-        service = Service(self, '/rosapi/service_response_details',
-                          'rosapi/ServiceResponseDetails')
-
-        service.call(ServiceRequest({'type': type}), callback, errback)
-
-    def get_nodes(self, callback, errback=None):
-        """Retrieve list of active node names in ROS."""
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+        """
         service = Service(self, '/rosapi/nodes',
                           'rosapi/Nodes')
 
-        service.call(ServiceRequest(), callback, errback)
+        result = service.call(ServiceRequest(), callback, errback)
 
-    def get_node_details(self, node, callback, errback=None):
-        """Retrieve list subscribed topics, publishing topics and services of a specific node name."""
+        if callback:
+            return
+
+        assert 'nodes' in result
+        return result['nodes']
+
+    def get_node_details(self, node, callback=None, errback=None):
+        """Retrieve list subscribed topics, publishing topics and services of a specific node name.
+
+        Note:
+            To make this a blocking call, pass ``None`` to the ``callback`` parameter .
+        """
         service = Service(self, '/rosapi/node_details',
                           'rosapi/NodeDetails')
 
-        service.call(ServiceRequest({'node': node}), callback, errback)
+        result = service.call(ServiceRequest({'node': node}), callback, errback)
+        if callback:
+            return
 
-    def get_params(self, callback, errback=None):
-        """Retrieve list of param names from the ROS Parameter Server."""
-        service = Service(self, '/rosapi/get_param_names',
-                          'rosapi/GetParamNames')
+        output = {
+            'services': result['services'],
+            'subscribing':  result['subscribing'],
+            'publishing': result['publishing']
+        }
 
-        service.call(ServiceRequest(), callback, errback)
-
-    def get_topic_type(self, topic, callback, errback=None):
-        """Retrieve the type of a topic in ROS."""
-        service = Service(self, '/rosapi/topic_type',
-                          'rosapi/TopicType')
-
-        service.call(ServiceRequest({'topic': topic}), callback, errback)
-
-    def get_service_type(self, service_name, callback, errback=None):
-        """Retrieve the type of a service in ROS."""
-        service = Service(self, '/rosapi/service_type',
-                          'rosapi/ServiceType')
-
-        service.call(ServiceRequest(
-            {'service': service_name}), callback, errback)
-
-    def get_message_details(self, message_type, callback, errback=None):
-        """Retrieve details of a message type in ROS."""
-        service = Service(self, '/rosapi/message_details',
-                          'rosapi/MessageDetails')
-
-        service.call(ServiceRequest(
-            {'type': message_type}), callback, errback)
+        return output
 
 
 if __name__ == '__main__':